# components/onboarding/api/swagger.yaml
basePath: /
definitions:
...
  AccountType:
    description: AccountType object
    properties:
      createdAt:
        description: The timestamp when the account type was created.
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      deletedAt:
        description: The timestamp when the account type was deleted.
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      description:
        description: Detailed description of the account type.
        example: Assets that are expected to be converted to cash within one year
        type: string
      id:
        description: The unique identifier of the Account Type.
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        type: string
      keyValue:
        description: A unique key value identifier for the account type.
        example: current_assets
        type: string
      ledgerId:
        description: The unique identifier of the Ledger.
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        type: string
      metadata:
        additionalProperties: {}
        description: |-
          Custom key-value pairs for extending the account type information
          example: {"department": "Treasury", "purpose": "Operating Expenses", "region": "Global"}
        type: object
      name:
        description: The name of the account type.
        example: Current Assets
        type: string
      organizationId:
        description: The unique identifier of the Organization.
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        type: string
      updatedAt:
        description: The timestamp when the account type was last updated.
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
  CreateAccountTypeInput:
    description: CreateAccountTypeInput payload
    properties:
      description:
        description: Detailed description of the account type.
        example: Assets that are expected to be converted to cash within one year
        maxLength: 500
        type: string
      keyValue:
        description: A unique key value identifier for the account type.
        example: current_assets
        maxLength: 50
        type: string
      metadata:
        additionalProperties: {}
        description: |-
          Custom key-value pairs for extending the account type information
          required: false
          example: {"department": "Treasury", "purpose": "Operating Expenses", "region": "Global"}
        type: object
      name:
        description: The name of the account type.
        example: Current Assets
        maxLength: 100
        type: string
    required:
    - keyValue
    - name
    type: object
...
  Error:
    description: Standardized error response format used across all API endpoints
      for error situations. Provides structured information about errors including
      codes, messages, and field-specific validation details.
    properties:
      code:
        description: |-
          Error code identifying the specific error condition
          example: ERR_INVALID_INPUT
          maxLength: 50
        example: ERR_INVALID_INPUT
        maxLength: 50
        type: string
      entityType:
        description: |-
          Optional type of entity associated with the error
          example: Organization
          maxLength: 100
        example: Organization
        maxLength: 100
        type: string
      fields:
        additionalProperties:
          type: string
        description: |-
          Optional detailed field validations for client-side handling
          example: {"name": "Field 'name' is required"}
        example:
          '{"name"': ' "Field ''name'' is required"}'
        type: object
      message:
        description: |-
          Detailed error message explaining the issue
          example: The request contains invalid fields. Please check the field 'name' and try again.
          maxLength: 500
        example: The request contains invalid fields. Please check the field 'name'
          and try again.
        maxLength: 500
        type: string
      title:
        description: |-
          Short, human-readable error title
          example: Bad Request
          maxLength: 100
        example: Bad Request
        maxLength: 100
        type: string
    type: object
...
  Pagination:
    description: Pagination is the struct designed to store the pagination data of
      an entity list.
    properties:
      items: {}
      limit:
        example: 10
        type: integer
      next_cursor:
        example: MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAwMA==
        type: string
        x-omitempty: true
      page:
        example: 1
        type: integer
      prev_cursor:
        example: MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAwMA==
        type: string
        x-omitempty: true
    type: object
...
  UpdateAccountTypeInput:
    description: UpdateAccountTypeInput payload
    properties:
      description:
        description: Detailed description of the account type.
        example: Assets that are expected to be converted to cash within one year
        maxLength: 500
        type: string
      metadata:
        additionalProperties: {}
        description: |-
          Custom key-value pairs for extending the account type information
          required: false
          example: {"department": "Treasury", "purpose": "Operating Expenses", "region": "Global"}
        type: object
      name:
        description: The name of the account type.
        example: Current Assets
        maxLength: 100
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    name: Discord community
    url: https://discord.gg/DnhqKwkGv3
  description: This is a swagger documentation for the Midaz Ledger API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Midaz Onboarding API
  version: v1.48.0
paths:
...
  /v1/organizations/{organization_id}/ledgers/{ledger_id}/account-types:
    get:
      description: Returns a paginated list of all account types for the specified
        organization and ledger, optionally filtered by metadata
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: JSON string to filter account types by metadata fields
        in: query
        name: metadata
        type: string
      - description: 'Limit of account types per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Page number for offset pagination (default: 1)'
        in: query
        name: page
        type: integer
      - description: Cursor for cursor-based pagination
        in: query
        name: cursor
        type: string
      - description: 'Sort order (asc or desc, default: asc)'
        in: query
        name: sort_order
        type: string
      - description: Start date for filtering (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date for filtering (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved account types
          schema:
            allOf:
            - $ref: '#/definitions/Pagination'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/AccountType'
                  type: array
                limit:
                  type: integer
                next_cursor:
                  type: string
                page:
                  type: object
                prev_cursor:
                  type: string
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Organization, ledger, or account types not found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Get all account types
      tags:
      - Account Types
    post:
      consumes:
      - application/json
      description: Endpoint to create a new Account Type.
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Account Type Input
        in: body
        name: accountType
        required: true
        schema:
          $ref: '#/definitions/CreateAccountTypeInput'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created account type
          schema:
            $ref: '#/definitions/AccountType'
        "400":
          description: Invalid input, validation errors
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "409":
          description: Conflict - account type key value already exists
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Create Account Type
      tags:
      - Account Types
  /v1/organizations/{organization_id}/ledgers/{ledger_id}/account-types/{id}:
    delete:
      description: Deletes an existing account type identified by its UUID within
        the specified ledger
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Account Type ID in UUID format
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted account type
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Account type not found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Delete an account type
      tags:
      - Account Types
    get:
      description: Returns detailed information about an account type identified by
        its UUID within the specified ledger
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Account Type ID in UUID format
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved account type
          schema:
            $ref: '#/definitions/AccountType'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Account type, ledger, or organization not found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Retrieve a specific account type
      tags:
      - Account Types
    patch:
      consumes:
      - application/json
      description: Endpoint to update an existing Account Type.
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Account Type ID in UUID format
        in: path
        name: id
        required: true
        type: string
      - description: Account Type Update Input
        in: body
        name: accountType
        required: true
        schema:
          $ref: '#/definitions/UpdateAccountTypeInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated account type
          schema:
            $ref: '#/definitions/AccountType'
        "400":
          description: Invalid input, validation errors
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Account type not found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Update Account Type
      tags:
      - Account Types
...
swagger: "2.0"